requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==8.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat
flower==0.9.7  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==3.0.11  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.11.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
django-dynamic-formsets==0.0.8

